Add program to enable forwarding of ports 0x910/0x911
On the Latitude E6430, these ports aren't forwarded over LPC when the
host OS is running. It seems like SMM enables these for itself before
disabling them again when it returns control to the OS. GEN2_DEC is set
up to forward ports 0x920 to 0x93f and 0x960 to 0x97f over LPC, so
changing it to forwarding ports 0x900 to 0x97f should be safe.